moved tests to match src file structure
