+++ tests/unit/cli/commands/available.test.ts
@@ -1,9 +1,9 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { availableCommand } from '../../src/cli/commands/available.js';
-import { ConfigDiscovery } from '../../src/core/config-discovery.js';
-import { MockSystemInterface } from '../mocks/mock-system-interface.js';
-import { createEnhancedMockFileSystem } from '../helpers/file-system-helpers.js';
+import { availableCommand } from '../../../../src/cli/commands/available.js';
+import { ConfigDiscovery } from '../../../../src/core/config-discovery.js';
+import { MockSystemInterface } from '../../mocks/mock-system-interface.js';
+import { createEnhancedMockFileSystem } from '../../helpers/file-system-helpers.js';
 
 // Mock dependencies
 jest.mock('fs');
