+++ tests/unit/cli/commands/format.test.ts
@@ -1,11 +1,11 @@
 import * as path from 'path';
-import { formatCommand, formatAllCommand } from '../../src/cli/commands/format.js';
-import { ConfigDiscovery } from '../../src/core/config-discovery.js';
-import { WorkflowEngine } from '../../src/core/workflow-engine.js';
+import { formatCommand, formatAllCommand } from '../../../../src/cli/commands/format.js';
+import { ConfigDiscovery } from '../../../../src/core/config-discovery.js';
+import { WorkflowEngine } from '../../../../src/core/workflow-engine.js';
 
 // Mock dependencies
 jest.mock('path');
-jest.mock('../../src/core/workflow-engine.js');
+jest.mock('../../../../src/core/workflow-engine.js');
 
 const mockPath = path as jest.Mocked<typeof path>;
 const MockedWorkflowEngine = WorkflowEngine as jest.MockedClass<typeof WorkflowEngine>;
