+++ tests/unit/cli/shared/cli-base.test.ts
@@ -1,17 +1,17 @@
-import { ConfigDiscovery } from '../../../src/core/config-discovery.js';
-import { WorkflowEngine } from '../../../src/core/workflow-engine.js';
-import type { Collection } from '../../../src/core/types.js';
+import { ConfigDiscovery } from '../../../../src/core/config-discovery.js';
+import { WorkflowEngine } from '../../../../src/core/workflow-engine.js';
+import type { Collection } from '../../../../src/core/types.js';
 import {
   initializeProject,
   initializeWorkflowEngine,
   validateWorkflow,
   validateCollection,
   findCollectionPath,
-} from '../../../src/cli/shared/cli-base.js';
+} from '../../../../src/cli/shared/cli-base.js';
 
 // Mock dependencies
-jest.mock('../../../src/core/config-discovery.js');
-jest.mock('../../../src/core/workflow-engine.js');
+jest.mock('../../../../src/core/config-discovery.js');
+jest.mock('../../../../src/core/workflow-engine.js');
 
 const MockConfigDiscovery = ConfigDiscovery as jest.MockedClass<typeof ConfigDiscovery>;
 const MockWorkflowEngine = WorkflowEngine as jest.MockedClass<typeof WorkflowEngine>;
