+++ tests/unit/cli/shared/workflow-operations.test.ts
@@ -5,9 +5,9 @@ import {
   loadWorkflowDefinition,
   scrapeUrlForCollection,
   findCollectionPath,
-} from '../../../src/cli/shared/workflow-operations.js';
-import { WorkflowFileSchema, type WorkflowFile } from '../../../src/core/schemas.js';
-import { scrapeUrl } from '../../../src/shared/web-scraper.js';
+} from '../../../../src/cli/shared/workflow-operations.js';
+import { WorkflowFileSchema, type WorkflowFile } from '../../../../src/core/schemas.js';
+import { scrapeUrl } from '../../../../src/shared/web-scraper.js';
 
 type WorkflowDefinition = {
   workflow: {
@@ -25,7 +25,7 @@ type WorkflowDefinition = {
 jest.mock('fs');
 jest.mock('path');
 jest.mock('yaml');
-jest.mock('../../../src/shared/web-scraper.js');
+jest.mock('../../../../src/shared/web-scraper.js');
 
 const mockFs = fs as jest.Mocked<typeof fs>;
 const mockPath = path as jest.Mocked<typeof path>;
